use uvarint

+++ ugo/utils/streamframe_interval.go
@@ -3,6 +3,6 @@ package utils
 // ByteInterval is an interval from one ByteCount to the other
 // +gen linkedlist
 type ByteInterval struct {
-	Start uint32
-	End   uint32
+	Start uint64
+	End   uint64
 }
